+++ sdk/typescript-schema/common/model.ts
@@ -1,3 +1,0 @@
-type AllModels = "all";
-export type Model = { type: string, id: number, name?: string };
-export type Models = AllModels | Model[];