+++ sdk/typescript-schema/CHANGELOG.md
@@ -1,3 +1,53 @@
+## [1.3.3](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.3.2...audience-typescript-schema@1.3.3) (2025-07-22)
+
+
+### Bug Fixes
+
+* don't use modules when re-compiling version ([#39](https://github.com/mParticle/audience-sdk/issues/39)) ([6f7d691](https://github.com/mParticle/audience-sdk/commit/6f7d691ff806c1d5665c4fd1bd58e07faf55df76))
+
+## [1.3.2](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.3.1...audience-typescript-schema@1.3.2) (2025-07-22)
+
+## [1.3.1](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.3.0...audience-typescript-schema@1.3.1) (2025-07-21)
+
+
+### Bug Fixes
+
+* emit js files on build ([#37](https://github.com/mParticle/audience-sdk/issues/37)) ([a7b76a0](https://github.com/mParticle/audience-sdk/commit/a7b76a06320ad2e87009e76a1b6c9f852a740f72))
+
+# [1.3.0](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.2.1...audience-typescript-schema@1.3.0) (2025-07-18)
+
+
+### Bug Fixes
+
+* Add audience operators to binary operator ([#36](https://github.com/mParticle/audience-sdk/issues/36)) ([3749153](https://github.com/mParticle/audience-sdk/commit/37491534a1a202e74346592be4ec25cc5a433148))
+
+
+### Features
+
+* Add in audience operator ([#33](https://github.com/mParticle/audience-sdk/issues/33)) ([ca98057](https://github.com/mParticle/audience-sdk/commit/ca980574f8c087e4f68b2ce2656a453a34967fce))
+* Change audience ID type from number -> string ([#35](https://github.com/mParticle/audience-sdk/issues/35)) ([5297227](https://github.com/mParticle/audience-sdk/commit/52972273beca7593441915b10c35c7f1ce3537eb))
+
+## [1.2.1](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.2.0...audience-typescript-schema@1.2.1) (2025-06-06)
+
+# [1.2.0](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.13...audience-typescript-schema@1.2.0) (2025-06-04)
+
+
+### Bug Fixes
+
+* prepare script ([a543a05](https://github.com/mParticle/audience-sdk/commit/a543a059e8c5947fc69520fc5ec0fd2cf5e9c29c))
+
+
+### Features
+
+* guards for audience/expression ([#25](https://github.com/mParticle/audience-sdk/issues/25)) ([947ee8c](https://github.com/mParticle/audience-sdk/commit/947ee8c1dbfcccf5750263dd508069e7bff3cd9c))
+* update field names for aggregator operands ([#24](https://github.com/mParticle/audience-sdk/issues/24)) ([47fe95f](https://github.com/mParticle/audience-sdk/commit/47fe95f0efd9ba43cc86fb9689e6388eb96245e2))
+
+## [1.1.13](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.12...audience-typescript-schema@1.1.13) (2025-06-02)
+
+## [1.1.12](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.11...audience-typescript-schema@1.1.12) (2025-05-22)
+
+## [1.1.11](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.10...audience-typescript-schema@1.1.11) (2025-05-16)
+
 ## [1.1.10](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.9...audience-typescript-schema@1.1.10) (2025-04-28)
 
 ## [1.1.9](https://github.com/mParticle/audience-sdk/compare/audience-typescript-schema@1.1.8...audience-typescript-schema@1.1.9) (2025-04-28)